chore(ui_adapter): split into modules
