+++ ui/src/warp_runner/ui_adapter/mod.rs
@@ -2,52 +2,23 @@
 //! a translation must be performed by WarpRunner.
 //!
 
+mod message_event;
+mod multipass_event;
+mod raygun_event;
+
+pub use message_event::{convert_message_event, MessageEvent};
+pub use multipass_event::{convert_multipass_event, MultiPassEvent};
+pub use raygun_event::{convert_raygun_event, RayGunEvent};
+
 use std::collections::VecDeque;
 
-use uuid::Uuid;
 use warp::{
     crypto::DID,
     error::Error,
-    multipass::MultiPassEventKind,
-    raygun::{self, Conversation, MessageEventKind, MessageOptions, RayGunEventKind},
+    raygun::{self, Conversation, MessageOptions},
 };
 
-use crate::{
-    logger,
-    state::{self, chats},
-};
-
-use super::conv_stream;
-
-#[allow(clippy::large_enum_variant)]
-pub enum RayGunEvent {
-    ConversationCreated(state::Chat),
-    ConversationDeleted(Uuid),
-}
-
-pub enum MultiPassEvent {
-    None,
-    FriendRequestReceived(state::Identity),
-    FriendRequestSent(state::Identity),
-    FriendAdded(state::Identity),
-    FriendRemoved(state::Identity),
-    FriendRequestCancelled(state::Identity),
-    FriendOnline(state::Identity),
-    FriendOffline(state::Identity),
-    Blocked(state::Identity),
-    Unblocked(state::Identity),
-}
-
-pub enum MessageEvent {
-    Received {
-        conversation_id: Uuid,
-        message: raygun::Message,
-    },
-    Sent {
-        conversation_id: Uuid,
-        message: raygun::Message,
-    },
-}
+use crate::state::{self, chats};
 
 pub async fn did_to_identity(
     did: &DID,
@@ -105,105 +76,3 @@ pub async fn conversation_to_chat(
 }
 
 // todo: put account and messaging in a module
-pub async fn convert_multipass_event(
-    event: warp::multipass::MultiPassEventKind,
-    account: &mut super::Account,
-    _messaging: &mut super::Messaging,
-) -> Result<MultiPassEvent, Error> {
-    //println!("got {:?}", &event);
-    let evt = match event {
-        MultiPassEventKind::FriendRequestSent { to } => {
-            let identity = did_to_identity(&to, account).await?;
-            MultiPassEvent::FriendRequestSent(identity)
-        }
-        MultiPassEventKind::FriendRequestReceived { from } => {
-            let identity = did_to_identity(&from, account).await?;
-            //println!("friend request received: {:#?}", identity);
-            MultiPassEvent::FriendRequestReceived(identity)
-        }
-        MultiPassEventKind::IncomingFriendRequestClosed { did }
-        | MultiPassEventKind::IncomingFriendRequestRejected { did }
-        | MultiPassEventKind::OutgoingFriendRequestClosed { did }
-        | MultiPassEventKind::OutgoingFriendRequestRejected { did } => {
-            let identity = did_to_identity(&did, account).await?;
-            MultiPassEvent::FriendRequestCancelled(identity)
-        }
-        MultiPassEventKind::FriendAdded { did } => {
-            let identity = did_to_identity(&did, account).await?;
-            MultiPassEvent::FriendAdded(identity)
-        }
-        MultiPassEventKind::FriendRemoved { did } => {
-            let identity = did_to_identity(&did, account).await?;
-            MultiPassEvent::FriendRemoved(identity)
-        }
-        MultiPassEventKind::IdentityOnline { did } => {
-            let identity = did_to_identity(&did, account).await?;
-            MultiPassEvent::FriendOnline(identity)
-        }
-        MultiPassEventKind::IdentityOffline { did } => {
-            let identity = did_to_identity(&did, account).await?;
-            MultiPassEvent::FriendOffline(identity)
-        }
-    };
-
-    Ok(evt)
-}
-
-pub async fn convert_raygun_event(
-    event: warp::raygun::RayGunEventKind,
-    stream_manager: &mut conv_stream::Manager,
-    account: &mut super::Account,
-    messaging: &mut super::Messaging,
-) -> Result<RayGunEvent, Error> {
-    logger::debug(&format!("got {:?}", &event));
-    let evt = match event {
-        RayGunEventKind::ConversationCreated { conversation_id } => {
-            let conv = messaging.get_conversation(conversation_id).await?;
-            let chat = conversation_to_chat(&conv, account, messaging).await?;
-            stream_manager.add_stream(chat.id, messaging).await?;
-            RayGunEvent::ConversationCreated(chat)
-        }
-        RayGunEventKind::ConversationDeleted { conversation_id } => {
-            stream_manager.remove_stream(conversation_id);
-            RayGunEvent::ConversationDeleted(conversation_id)
-        }
-    };
-
-    Ok(evt)
-}
-
-pub async fn convert_message_event(
-    event: warp::raygun::MessageEventKind,
-    _account: &mut super::Account,
-    messaging: &mut super::Messaging,
-) -> Result<MessageEvent, Error> {
-    logger::debug(&format!("got {:?}", &event));
-    let evt = match event {
-        MessageEventKind::MessageReceived {
-            conversation_id,
-            message_id,
-        } => {
-            let message = messaging.get_message(conversation_id, message_id).await?;
-            MessageEvent::Received {
-                conversation_id,
-                message,
-            }
-        }
-        MessageEventKind::MessageSent {
-            conversation_id,
-            message_id,
-        } => {
-            let message = messaging.get_message(conversation_id, message_id).await?;
-            MessageEvent::Sent {
-                conversation_id,
-                message,
-            }
-        }
-        _ => {
-            println!("evt received: {:?}", event);
-            todo!();
-        }
-    };
-
-    Ok(evt)
-}
